Revert "Moved password/active changing to auth service."
This reverts commit 14982d50e0733a75912929eebc390ce92ec57e65.
