+++ Authentication/src/main/java/com/epam/wca/authentication/controller/AuthenticationController.java
@@ -1,6 +1,5 @@
 package com.epam.wca.authentication.controller;
 
-import com.epam.wca.authentication.dto.UserActivationDTO;
 import com.epam.wca.authentication.dto.UserUpdateDTO;
 import com.epam.wca.authentication.entity.User;
 import com.epam.wca.common.gymcommon.auth_dto.UserAuthenticatedDTO;
@@ -11,7 +10,6 @@ import org.springframework.http.MediaType;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -41,12 +39,6 @@ public interface AuthenticationController {
     @PutMapping(value = "/password", consumes = MediaType.APPLICATION_JSON_VALUE)
     String changePassword(
             @RequestBody @Valid UserUpdateDTO userDTO,
-            @AuthenticationPrincipal User authenticatedUser
-    );
-
-    @PatchMapping(value = "/active")
-    String activateDeactivate(
-            @RequestBody @Valid UserActivationDTO userDTO,
             @AuthenticationPrincipal User authenticatedUser
     );
 }
