+++ GymSpringCore/src/main/java/com/epam/wca/gym/entity/Trainer.java
@@ -48,7 +48,8 @@ public class Trainer extends User {
         super(
                 firstName,
                 lastName,
-                username
+                username,
+                true
         );
 
         this.specialization = specialization;
