+++ GymSpringCore/src/main/java/com/epam/wca/gym/service/UserService.java
@@ -1,7 +1,10 @@
 package com.epam.wca.gym.service;
 
+import com.epam.wca.gym.dto.user.UserActivationDTO;
 import com.epam.wca.gym.entity.User;
 
 public interface UserService {
+    void update(User user, UserActivationDTO userDTO);
+
     User findByUsername(String username);
 }
