+++ GymSpringCore/src/main/resources/sql/data.sql
@@ -1,37 +1,37 @@
 -- Populating the USERS table (30 users in total, 15 trainees and 15 trainers)
-INSERT INTO USERS (ID, FIRSTNAME, LASTNAME, USERNAME)
+INSERT INTO USERS (ID, FIRSTNAME, LASTNAME, USERNAME, IS_ACTIVE)
 VALUES
-    (1, 'John', 'Doe', 'john.doe'),
-    (2, 'Jane', 'Smith', 'jane.smith'),
-    (3, 'Alice', 'Brown', 'alice.brown'),
-    (4, 'Michael', 'Johnson', 'michael.johnson'),
-    (5, 'Emily', 'Davis', 'emily.davis'),
-    (6, 'David', 'Garcia', 'david.garcia'),
-    (7, 'Emma', 'Martinez', 'emma.martinez'),
-    (8, 'Oliver', 'Lopez', 'oliver.lopez'),
-    (9, 'Sophia', 'Gonzalez', 'sophia.gonzalez'),
-    (10, 'James', 'Wilson', 'james.wilson'),
-    (11, 'Charlotte', 'Moore', 'charlotte.moore'),
-    (12, 'Benjamin', 'Taylor', 'benjamin.taylor'),
-    (13, 'Lucas', 'Anderson', 'lucas.anderson'),
-    (14, 'Amelia', 'Thomas', 'amelia.thomas'),
-    (15, 'Mason', 'Jackson', 'mason.jackson'),
+    (1, 'John', 'Doe', 'john.doe', true),
+    (2, 'Jane', 'Smith', 'jane.smith', true),
+    (3, 'Alice', 'Brown', 'alice.brown', true),
+    (4, 'Michael', 'Johnson', 'michael.johnson', true),
+    (5, 'Emily', 'Davis', 'emily.davis', true),
+    (6, 'David', 'Garcia', 'david.garcia', true),
+    (7, 'Emma', 'Martinez', 'emma.martinez', true),
+    (8, 'Oliver', 'Lopez', 'oliver.lopez', true),
+    (9, 'Sophia', 'Gonzalez', 'sophia.gonzalez', true),
+    (10, 'James', 'Wilson', 'james.wilson', true),
+    (11, 'Charlotte', 'Moore', 'charlotte.moore', true),
+    (12, 'Benjamin', 'Taylor', 'benjamin.taylor', true),
+    (13, 'Lucas', 'Anderson', 'lucas.anderson', true),
+    (14, 'Amelia', 'Thomas', 'amelia.thomas', true),
+    (15, 'Mason', 'Jackson', 'mason.jackson', true),
 -- Trainers
-    (16, 'Ethan', 'White', 'ethan.white'),
-    (17, 'Olivia', 'Harris', 'olivia.harris'),
-    (18, 'Isabella', 'Clark', 'isabella.clark'),
-    (19, 'Aiden', 'Lewis', 'aiden.lewis'),
-    (20, 'Ava', 'Young', 'ava.young'),
-    (21, 'Jacob', 'Walker', 'jacob.walker'),
-    (22, 'Liam', 'Hall', 'liam.hall'),
-    (23, 'Mia', 'Allen', 'mia.allen'),
-    (24, 'Noah', 'King', 'noah.king'),
-    (25, 'Ella', 'Wright', 'ella.wright'),
-    (26, 'William', 'Scott', 'william.scott'),
-    (27, 'Abigail', 'Torres', 'abigail.torres'),
-    (28, 'Elijah', 'Nguyen', 'elijah.nguyen'),
-    (29, 'Avery', 'Hill', 'avery.hill'),
-    (30, 'Evelyn', 'Green', 'evelyn.green');
+    (16, 'Ethan', 'White', 'ethan.white', true),
+    (17, 'Olivia', 'Harris', 'olivia.harris', true),
+    (18, 'Isabella', 'Clark', 'isabella.clark', true),
+    (19, 'Aiden', 'Lewis', 'aiden.lewis', true),
+    (20, 'Ava', 'Young', 'ava.young', true),
+    (21, 'Jacob', 'Walker', 'jacob.walker', true),
+    (22, 'Liam', 'Hall', 'liam.hall', true),
+    (23, 'Mia', 'Allen', 'mia.allen', true),
+    (24, 'Noah', 'King', 'noah.king', true),
+    (25, 'Ella', 'Wright', 'ella.wright', true),
+    (26, 'William', 'Scott', 'william.scott', true),
+    (27, 'Abigail', 'Torres', 'abigail.torres', true),
+    (28, 'Elijah', 'Nguyen', 'elijah.nguyen', true),
+    (29, 'Avery', 'Hill', 'avery.hill', true),
+    (30, 'Evelyn', 'Green', 'evelyn.green', true);
 
 -- Populating the TRAINING_TYPE table
 INSERT INTO TRAINING_TYPE (TYPE)
